requirements.txt
PyPDF2
reportlab
pdfminer.six
